+++ src/test/java/baseEntities/BaseTest.java
@@ -5,6 +5,7 @@ import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import services.BrowsersService;
+import steps.OrderStep;
 import steps.UserStep;
 
 public class BaseTest {
@@ -13,11 +14,15 @@ public class BaseTest {
 
     protected UserStep userStep;
 
+    protected OrderStep orderStep;
+
     @BeforeMethod
     public void setUp() {
         driver = new BrowsersService().getDriver();
 
         userStep = new UserStep(driver);
+        orderStep = new OrderStep(driver);
+
 
         driver.get(ReadProperties.getUrl());
     }
