+++ src/test/java/pages/CartOfGoodsPage.java
@@ -12,7 +12,7 @@ public class CartOfGoodsPage extends BasePage {
     private final By TITLE_LABEL = By.xpath("//span[text() ='Your Cart']");
     private final By CONTINUE_SHOPPING_BUTTON = By.id("continue-shopping");
     private final By CHECKOUT_BUTTON = By.id("checkout");
-    private final By ITEM_IN_CART = By.xpath("//div[@Class ='inventory_item_name']");
+    private final By ITEM_IN_CART = By.xpath("//div[@class ='inventory_item_name']");
     private final By DELETE_ITEM_IN_CART_BUTTON = By.id("remove-sauce-labs-backpack");
