+++ src/test/java/tests/PaymentForGoodsTest.java
@@ -9,9 +9,9 @@ public class PaymentForGoodsTest extends BaseTest {
     @Test
     public void payment() {
         userStep.successfulLogin(ReadProperties.getUsername(), ReadProperties.getPassword());
-        userStep.addingItemToCart();
-        userStep.enteringPersonalData();
-        userStep.purchaseOfGoods();
-        userStep.finalCheck();
+        orderStep.addingItemToCart();
+        orderStep.enteringPersonalData();
+        orderStep.purchaseOfGoods();
+        orderStep.finalCheck();
     }
 }
